renamed page step components
